Fix the reference frame rate; make every effect own its background
Effects calibrated against 60 fps ran about 4% fast, because the shared frame
timer derived its reference period as 1000/60 and got 16 ms — a 62.5 Hz clock.
Separately, two effects painted only the pixels they lit, so they inherited
whatever was on screen: their own trails as permanent ghosts, and the previous
effect's whole picture as background.

desktop 140us/7,142fps | esp32 4,164us/240fps | flash: desktop 1,075KB,
esp32 1,704KB, esp32s3-n16r8 1,699KB

Core
- FrameTime carries the undivided numerator and divides late, the rule BeatPhase
  already follows. One unit is 1000/(256*60) = 0.065 ms, so a remainder held in
  whole milliseconds cannot represent it: the truncated time was discarded, and
  because the amount discarded differs by render rate (11 units a frame at 60
  fps, under 1 at 200) it was itself a framerate dependency. A simulated second
  now measures exactly 60 reference frames at 60 fps.
- atan16 folds through unsigned magnitudes. Negating INT32_MIN has no int32
  representation, so the previous fold was undefined behaviour at one input per
  axis — reachable from any unclamped coordinate difference. Confirmed by
  UBSan, which now runs clean over the range.

Light domain
- SdfShapes and Tetrix write their background instead of skipping it. The Layer
  does not clear between frames (ADR-0003), so an effect that only writes lit
  pixels shows the previous effect's frame around its own. Tetrix predates this
  branch.
- shader::project rejects a projection that does not fit int32 rather than
  truncating it, which wrapped geometry off one edge of the panel to the other.
- Truchet's floorDiv adjusts the quotient rather than negating the dividend, and
  Rings keeps its coordinate deltas in int32 — the same INT32_MIN class as
  atan16.
- Fireworks launches at launchRate 1-3. Its own control minimum is 1, and the
  rate divided by 4 truncated those to zero, so the three slowest settings the
  UI offers never launched anything.
- BouncingBalls carries its trail fade instead of rounding up to 1, which at
  high fps applied many times the intended decay.
- VectorBalls uses shader::rotate rather than its own copy of it.

Tests
- A new sweep ticks all 51 effects against a DIRTY buffer, which is what a
  device hands an effect on every frame after the first. Golden tests render
  into a zeroed buffer, so the pixels an effect never writes are black by luck
  and hash correctly — this is the shape of bug they cannot see. It found the
  Tetrix case on its first run. DemoReel and NetworkReceive are exempt with
  reasons: one delegates its frame to a child, the other blocks in recvfrom.
- Coverage for the exact reference rate, atan16 at INT32_MIN, and project at
  the limits of its range.
- Seven goldens recaptured: the reference-rate fix changes what every FrameTime
  user draws on a given frame, and makes the motion per second correct rather
  than merely consistent.
- BouncingBalls is recorded at a 1.40 framerate band with the reason in the
  test, rather than widening the band for all 51. Its motion is time-driven and
  correct; its trail rendering is what drifts, and it is the effect still to
  move onto the particle kernel.

Docs/CI
- architecture.md states why the frame timer divides late.
- lessons.md: four lessons from this branch — verifying a ported function's sign
  against upstream, asserting on a distribution rather than eyeballing it,
  compensating bugs that cancel, and proving a "no-op" primitive rewrite by
  exhaustive comparison.

Reviews
- Reviewer (Fable) over the branch diff, 4 fixed. Skipped: moving FrameTime to
  core and the Buffer-form draw API removal (both real, both refactors across
  every caller); the uncalled power-function surface (the planned MoonLive
  builtin set); em-dash prose sweep (opportunistic per coding-standards).
- CodeRabbit, 8 fixed. Skipped: lerp16's 65535 denominator, since t is a
  fraction of 65536 and changing it would bias every interpolation to correct
  one endpoint.

+++ src/core/math16.h
@@ -231,13 +231,16 @@ inline constexpr int16_t atan16_octant[33] = {
 inline angle16 atan16(int32_t y, int32_t x) {
     if (x == 0 && y == 0) return 0;                       // the centre has no direction
     // Fold into the first octant, remembering which one, then interpolate the arctangent there.
-    int32_t ax = x < 0 ? -x : x;
-    int32_t ay = y < 0 ? -y : y;
+    // Unsigned magnitudes: negating INT32_MIN has no int32 representation, so the obvious
+    // `x < 0 ? -x : x` is undefined behaviour at exactly one input per axis. Widening first keeps
+    // the fold exact across the whole range.
+    uint32_t ax = x < 0 ? -static_cast<uint32_t>(x) : static_cast<uint32_t>(x);
+    uint32_t ay = y < 0 ? -static_cast<uint32_t>(y) : static_cast<uint32_t>(y);
     const bool swap = ay > ax;
-    if (swap) { const int32_t t = ax; ax = ay; ay = t; }
+    if (swap) { const uint32_t t = ax; ax = ay; ay = t; }
     // ratio = ay/ax in 0..65535; within one octant arctan is near-linear, so a linear read with a
     // small cubic correction is well inside a pixel of error at any grid size we drive.
-    const uint32_t ratio = static_cast<uint32_t>((static_cast<uint64_t>(ay) << 16) / (ax ? ax : 1));
+    const uint32_t ratio = static_cast<uint32_t>((static_cast<uint64_t>(ay) << 16) / (ax ? ax : 1u));
     // Table lookup with linear interpolation, the same shape as sin16 above and for the same reason:
     // a fitted polynomial was tried first and measured 9.6 degrees of error at the octant boundary,
     // where a 66-byte table is exact at every entry and closes precisely at 8192.
